БД музей/requirements.txt
"""
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
"""